system/vold: Fix the issue of not able to mount 32 GB SD card after formatting it from UI.
Cluster for sector option is set to 8 while formatting SD card.
System is taking more memory for checking FAT on 32 GB SD card
because of more number of clusters. System is killing the applications
due to low memory. Android framework is getting rebooted when the Launcher
application is killed and SD card is not able mount.

Fix the issue by suppressing the sectors per cluster options while calling
the format utility. Format utility calculate sectors per cluster based on
the SD card size and number of clusters are optimized.

Change-Id: I4839da647bcf8f6e6d8fab55fd85be1277fd5fc7
CRs-fixed: 248144