system/vold: Fix the issue of not able to format SD/MMC card and mount after format from UI on 8660
Minor number is hardcoded in the devpath string, send it as argument
to the format utility (newfs_msdos). It is causing to picking wrong
devpath.

Fix the issue by deriving devapth from the device node minor number.

Volume state is set to formatting before starting the format and set
it back to device_idle state when the format is successful. This state
needs to be set to device_idle in format fail case also.

Change-Id: Ie9fec966abfaa146da0a0b42e8be723052ddc9b4
CRs-fixed: 253632