system/vold: fix the issue of not handling block events with different major numbers
Vold is fail to handle the partitions of SD/MMC cards greater
than 15, because it is assumed that the major number same for
all the partitions.

Fix the issue by storing Major number along with the minor
number and use it while checking and mounting the file system
present on the SD/MMC card.

Change-Id: Ic451644acdcb65e09c98537fae416183d48dc174