system/vold: Fix dirty writeback and expire values for usb mass storage
init.qcom.rc file is over-writing the dirty_writeback and dirty_expiry timer values
to 60000 centisecs (600 seconds) from the default Linux kernel values in-order to have
less frequent sync requests generated which in-directly helps in the less frequent timer
wakeups which results in better battery power management when connected to the battery.

These values doesn't help much while connected to the wall charger or USB connection
as we always draw power from the USB and power measurement doesn't help while USB is
connected to the system. Above timer values then creates the problem while doing the
USB mass storage copy through the host and it causes the more data to get buffered with the
less amount of sync (in the default case we have to wait 600 seconds to get the next sync)
which causes the USB re-enumeration. This can be avoided by setting the optimal values while
the USB is connected to the system.

CRs-Fixed: 286266

Change-Id: I1b7be7cbab87c8bf071ccf69a9d2cf2f7c61bd86