system/vold: Fix dirty writeback and expire values for external mass storage
init.qcom.rc file is over-writing the dirty_writeback and dirty_expiry timer values
to 60000 centisecs (600 seconds) from the default Linux kernel values in-order to have
less frequent sync requests generated which in-directly helps in the less frequent timer
wakeups which results in better battery power management when connected to the battery.

Above timer values then creates the problem while doing the External mass storage copy and
it causes the more data to get buffered with the less amount of sync (in the default case
we have to wait 600 seconds to get the next sync).

CRs-Fixed: 292167

Change-Id: I21ef360e6d3e12dc711792090a4991b94367cc65